Resolved 'failed to detect EUI verion: '
I resolved an issue of receiving a netaddr.core.AddrFormatError while using get_lldp_neighbors and get_lldp_neighbors_details. It seems as though the mac addresses that were returned on line 323 would fail if the mac address that was parsed was in a format that the mac function couldn't decipher. So now instead of returning an error, it will return lldp_entry[k] as ' ' if it cannot parse the mac address.