Fixed blank LLDP interface names in Netbox
After making the first changes I realized it now caused all the LLDP Interfaces to show blank. I reviewed my changes and I realized the issue was that the mac() function on line 324 was trying to parse an interface name "eth0" which is not a mac address so it caused the mac() function to throw the error I mentioned earlier. Now, at least for our use, it will return the name of the interface.